Extract BaseActivity class from TaskActivity.
